+++ lib/services/places_service.dart
@@ -1,6 +1,6 @@
 /// Service for managing places data in the user's Solid Pod.
 ///
-// Time-stamp: <2025-12-04 Miduo>
+// Time-stamp: <2026-01-07 Miduo>
 ///
 /// Copyright (C) 2025, Software Innovation Institute, ANU.
 ///
@@ -29,7 +29,6 @@ import 'dart:convert';
 
 import 'package:flutter/material.dart';
 
-import 'package:shared_preferences/shared_preferences.dart';
 import 'package:solidpod/solidpod.dart';
 
 import 'package:geopod/constants/example_places_data.dart';
@@ -37,6 +36,7 @@ import 'package:geopod/models/place.dart';
 import 'package:geopod/services/geocoding_service.dart';
 import 'package:geopod/services/places/encrypted_places_service.dart';
 import 'package:geopod/services/places/places_cache_manager.dart';
+import 'package:geopod/services/places/places_cache_persistence.dart';
 import 'package:geopod/services/places/places_import_export.dart';
 import 'package:geopod/services/places/places_pod_file.dart';
 import 'package:geopod/services/pod/pod_directory_service.dart';
@@ -46,10 +46,6 @@ export 'package:geopod/services/places/places_cache_manager.dart';
 export 'package:geopod/services/places/places_import_export.dart';
 
 final placesChangeNotifier = ValueNotifier<int>(0);
-
-const String _keyPodPlacesCache = 'pod_places_cache';
-const String _keyPodPlacesCacheTimestamp = 'pod_places_cache_timestamp';
-const Duration _cacheExpiry = Duration(minutes: 5);
 
 class PlacesService {
   /// Cached local places (lazily initialized from compiled constants).
@@ -125,7 +121,7 @@ class PlacesService {
         }
       }
       if (!forceRefresh) {
-        final c = await _getCachedPodPlaces();
+        final c = await PlacesCachePersistence.getCachedPodPlaces();
         if (c != null) {
           cm.cachePodPlaces(c);
           _refreshPodPlacesInBackground();
@@ -149,49 +145,10 @@ class PlacesService {
         } catch (_) {}
       }
       places.sort((a, b) => b.timestamp.compareTo(a.timestamp));
-      await _cachePodPlaces(content);
+      await PlacesCachePersistence.cachePodPlaces(content);
       cm.cachePodPlaces(places);
     } catch (_) {}
     return places;
-  }
-
-  static Future<List<Place>?> _getCachedPodPlaces() async {
-    try {
-      final prefs = await SharedPreferences.getInstance();
-      final json = prefs.getString(_keyPodPlacesCache);
-      final ts = prefs.getInt(_keyPodPlacesCacheTimestamp);
-      if (json == null || ts == null) return null;
-      if (DateTime.now().difference(DateTime.fromMillisecondsSinceEpoch(ts)) >
-          _cacheExpiry) {
-        return null;
-      }
-      final places = <Place>[];
-      final decoded = jsonDecode(json);
-      if (decoded is List) {
-        for (final i in decoded) {
-          if (i is Map<String, dynamic>) {
-            try {
-              places.add(Place.fromJson(i, isLocalSource: false));
-            } catch (_) {}
-          }
-        }
-      }
-      places.sort((a, b) => b.timestamp.compareTo(a.timestamp));
-      return places;
-    } catch (_) {
-      return null;
-    }
-  }
-
-  static Future<void> _cachePodPlaces(String json) async {
-    try {
-      final p = await SharedPreferences.getInstance();
-      await p.setString(_keyPodPlacesCache, json);
-      await p.setInt(
-        _keyPodPlacesCacheTimestamp,
-        DateTime.now().millisecondsSinceEpoch,
-      );
-    } catch (_) {}
   }
 
   static void _refreshPodPlacesInBackground() {
@@ -199,7 +156,9 @@ class PlacesService {
       try {
         if (!AuthDataManager.isLoggedInSync()) return;
         final c = await readPlacesJsonFile();
-        if (c != null && c.trim().isNotEmpty) await _cachePodPlaces(c);
+        if (c != null && c.trim().isNotEmpty) {
+          await PlacesCachePersistence.cachePodPlaces(c);
+        }
       } catch (_) {}
     });
   }
@@ -207,18 +166,14 @@ class PlacesService {
   static Future<void> clearCache() async {
     try {
       PlacesCacheManager().clearCache();
-      final p = await SharedPreferences.getInstance();
-      await p.remove(_keyPodPlacesCache);
-      await p.remove(_keyPodPlacesCacheTimestamp);
+      await PlacesCachePersistence.clearPodPlacesCache();
     } catch (_) {}
   }
 
   static Future<void> clearPodCacheOnly() async {
     try {
       PlacesCacheManager().clearPodCacheOnly();
-      final p = await SharedPreferences.getInstance();
-      await p.remove(_keyPodPlacesCache);
-      await p.remove(_keyPodPlacesCacheTimestamp);
+      await PlacesCachePersistence.clearPodPlacesCache();
     } catch (_) {}
   }
