+++ lib/widgets/map_settings_dialog.dart
@@ -1,6 +1,6 @@
 /// Dialog for configuring map display settings.
 ///
-// Time-stamp: <2025-12-05 Miduo>
+// Time-stamp: <2026-01-07 Miduo>
 ///
 /// Copyright (C) 2025, Software Innovation Institute, ANU.
 ///
@@ -29,19 +29,17 @@ import 'dart:async' show unawaited;
 
 import 'package:flutter/material.dart';
 
-import 'package:flutter_colorpicker/flutter_colorpicker.dart';
-import 'package:solidpod/solidpod.dart';
-import 'package:solidui/solidui.dart';
-
 import 'package:geopod/services/map_settings_service.dart';
-import 'package:geopod/widgets/settings/color_picker_tile.dart';
-import 'package:geopod/widgets/settings/viewport_selector.dart';
+import 'package:geopod/widgets/settings/settings_actions.dart';
+import 'package:geopod/widgets/settings/settings_sections.dart';
 
 /// Dialog for configuring map display settings.
 ///
 /// Allows users to:
 /// - Toggle visibility of local (canned) example places
 /// - Customize colors for user places and example places
+/// - Select map source
+/// - Configure viewport settings
 class MapSettingsDialog extends StatefulWidget {
   const MapSettingsDialog({
     super.key,
@@ -105,113 +103,6 @@ class _MapSettingsDialogState extends State<MapSettingsDialog> {
         _initialZoom != _initialSnapshot.initialZoom;
   }
 
-  /// DEBUG: Delete encryption keys from server
-  Future<void> _deleteEncryptionKeys(BuildContext context) async {
-    final confirmed = await showDialog<bool>(
-      context: context,
-      builder: (ctx) => AlertDialog(
-        title: const Text('Delete Encryption Keys'),
-        content: const Text(
-          'This will delete all encryption keys from the server.\n\n'
-          'This includes:\n'
-          '• enc-keys.ttl (verification key + private key)\n'
-          '• ind-keys.ttl (individual file keys)\n'
-          '• public-key.ttl (public key)\n\n'
-          'WARNING: All encrypted data will become unreadable!\n'
-          'This action cannot be undone.',
-        ),
-        actions: [
-          TextButton(
-            onPressed: () => Navigator.pop(ctx, false),
-            child: const Text('Cancel'),
-          ),
-          ElevatedButton(
-            onPressed: () => Navigator.pop(ctx, true),
-            style: ElevatedButton.styleFrom(
-              backgroundColor: Colors.red,
-              foregroundColor: Colors.white,
-            ),
-            child: const Text('Delete'),
-          ),
-        ],
-      ),
-    );
-
-    if (confirmed != true || !context.mounted) return;
-
-    try {
-      // Show loading indicator
-      showDialog(
-        context: context,
-        barrierDismissible: false,
-        builder: (_) => const Center(child: CircularProgressIndicator()),
-      );
-
-      // Get the webId to construct full URLs
-      final webId = await getWebId();
-      if (webId == null || webId.isEmpty) {
-        throw Exception('User not logged in');
-      }
-
-      // appDirName for geopod is 'geopod', files are relative to POD root
-      // deleteFile expects path relative to POD root (not appDirName)
-      // The file paths should be: geopod/encryption/enc-keys.ttl etc.
-      final filesToDelete = [
-        'geopod/encryption/enc-keys.ttl',
-        'geopod/encryption/ind-keys.ttl',
-        'geopod/sharing/public-key.ttl',
-      ];
-
-      final deletedFiles = <String>[];
-      final failedFiles = <String>[];
-
-      for (final filePath in filesToDelete) {
-        try {
-          debugPrint('Attempting to delete: $filePath');
-          // deleteFile with isKey=true only deletes the file without
-          // trying to revoke permissions or remove individual keys
-          await deleteFile(filePath, isKey: true);
-          deletedFiles.add(filePath);
-          debugPrint('Deleted: $filePath');
-        } catch (e) {
-          failedFiles.add(filePath);
-          debugPrint('Failed to delete $filePath: $e');
-        }
-      }
-
-      // Clear local key cache
-      await KeyManager.clear();
-
-      if (context.mounted) {
-        Navigator.pop(context); // Close loading
-        Navigator.pop(context); // Close settings dialog
-
-        final message = deletedFiles.isNotEmpty
-            ? 'Deleted ${deletedFiles.length} files. '
-                  '${failedFiles.isNotEmpty ? "Failed: ${failedFiles.length}" : ""}'
-                  '\nPlease logout and login.'
-            : 'No files were deleted. ${failedFiles.length} failures.';
-
-        ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(
-            content: Text(message),
-            backgroundColor: deletedFiles.isNotEmpty
-                ? Colors.orange
-                : Colors.red,
-          ),
-        );
-      }
-    } catch (e) {
-      debugPrint('Error deleting encryption keys: $e');
-      if (context.mounted) {
-        Navigator.pop(context); // Close loading
-        ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(content: Text('Error: $e'), backgroundColor: Colors.red),
-        );
-      }
-    }
-  }
-
   /// Saves current settings and notifies parent.
   void _saveAndNotify() {
     final newSettings = MapSettings(
@@ -232,64 +123,6 @@ class _MapSettingsDialogState extends State<MapSettingsDialog> {
 
     // Notify parent widget.
     widget.onSettingsChanged(newSettings);
-  }
-
-  /// Shows color picker dialog for selecting a color.
-  Future<void> _showColorPicker({
-    required String title,
-    required Color currentColor,
-    required void Function(Color) onColorChanged,
-  }) async {
-    Color selectedColor = currentColor;
-
-    await showDialog(
-      context: context,
-      builder: (context) => AlertDialog(
-        title: Text(title),
-        content: SingleChildScrollView(
-          child: BlockPicker(
-            pickerColor: currentColor,
-            onColorChanged: (color) {
-              selectedColor = color;
-            },
-            availableColors: const [
-              Colors.red,
-              Colors.pink,
-              Colors.purple,
-              Colors.deepPurple,
-              Colors.indigo,
-              Colors.blue,
-              Colors.lightBlue,
-              Colors.cyan,
-              Colors.teal,
-              Colors.green,
-              Colors.lightGreen,
-              Colors.lime,
-              Colors.yellow,
-              Colors.amber,
-              Colors.orange,
-              Colors.deepOrange,
-              Colors.brown,
-              Colors.grey,
-              Colors.blueGrey,
-            ],
-          ),
-        ),
-        actions: [
-          TextButton(
-            onPressed: () => Navigator.pop(context),
-            child: const Text('Cancel'),
-          ),
-          ElevatedButton(
-            onPressed: () {
-              onColorChanged(selectedColor);
-              Navigator.pop(context);
-            },
-            child: const Text('Select'),
-          ),
-        ],
-      ),
-    );
   }
 
   /// Resets all settings to defaults.
@@ -330,281 +163,87 @@ class _MapSettingsDialogState extends State<MapSettingsDialog> {
             mainAxisSize: MainAxisSize.min,
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
-              // Visibility toggle.
-              const Text(
-                'Visibility',
-                style: TextStyle(
-                  fontSize: 14,
-                  fontWeight: FontWeight.bold,
-                  color: Colors.grey,
-                ),
-              ),
-              const SizedBox(height: 8),
-              SwitchListTile(
-                title: const Text('Show Example Locations'),
-                subtitle: const Text('Display canned examples on map'),
-                value: _showLocalPlaces,
-                onChanged: (value) {
+              // Visibility section
+              buildVisibilitySection(
+                showLocalPlaces: _showLocalPlaces,
+                showEncryptedPlaces: _showEncryptedPlaces,
+                isLoadingEncrypted: _isLoadingEncrypted,
+                onShowLocalChanged: (value) {
                   setState(() => _showLocalPlaces = value);
                   _saveAndNotify();
                 },
-                secondary: Icon(
-                  _showLocalPlaces ? Icons.visibility : Icons.visibility_off,
-                  color: _showLocalPlaces ? Colors.green : Colors.grey,
-                ),
-              ),
-              const SizedBox(height: 8),
-              SwitchListTile(
-                title: const Text('Show Encrypted Places'),
-                subtitle: Text(
-                  _isLoadingEncrypted
-                      ? 'Loading encrypted data...'
-                      : 'Display encrypted places (requires key)',
-                ),
-                value: _showEncryptedPlaces,
-                onChanged: _isLoadingEncrypted
-                    ? null
-                    : (value) {
-                        setState(() {
-                          _showEncryptedPlaces = value;
-                          if (value) _isLoadingEncrypted = true;
-                        });
-                        _saveAndNotify();
-                        // Reset loading state after a delay
-                        if (value) {
-                          Future.delayed(const Duration(seconds: 3), () {
-                            if (mounted) {
-                              setState(() => _isLoadingEncrypted = false);
-                            }
-                          });
-                        }
-                      },
-                secondary: _isLoadingEncrypted
-                    ? const SizedBox(
-                        width: 24,
-                        height: 24,
-                        child: CircularProgressIndicator(strokeWidth: 2),
-                      )
-                    : Icon(
-                        _showEncryptedPlaces ? Icons.lock_open : Icons.lock,
-                        color: _showEncryptedPlaces
-                            ? Colors.purple
-                            : Colors.grey,
-                      ),
+                onShowEncryptedChanged: (value) {
+                  setState(() {
+                    _showEncryptedPlaces = value;
+                    if (value) _isLoadingEncrypted = true;
+                  });
+                  _saveAndNotify();
+                  // Reset loading state after a delay
+                  if (value) {
+                    Future.delayed(const Duration(seconds: 3), () {
+                      if (mounted) {
+                        setState(() => _isLoadingEncrypted = false);
+                      }
+                    });
+                  }
+                },
               ),
               const Divider(height: 24),
 
-              // Viewport Settings
-              const Text(
-                'Viewport',
-                style: TextStyle(
-                  fontSize: 14,
-                  fontWeight: FontWeight.bold,
-                  color: Colors.grey,
-                ),
-              ),
-              const SizedBox(height: 8),
-              SwitchListTile(
-                title: const Text('Remember Viewport'),
-                subtitle: const Text('Resume from last viewed position'),
-                value: _rememberViewport,
-                onChanged: (value) {
+              // Viewport section
+              buildViewportSection(
+                rememberViewport: _rememberViewport,
+                initialLat: _initialLat,
+                initialLng: _initialLng,
+                initialZoom: _initialZoom,
+                onRememberViewportChanged: (value) {
                   setState(() => _rememberViewport = value);
                   _saveAndNotify();
                 },
-                secondary: Icon(
-                  _rememberViewport ? Icons.restore : Icons.home,
-                  color: _rememberViewport ? Colors.blue : Colors.grey,
-                ),
+                onViewportChanged: (lat, lng, zoom) {
+                  setState(() {
+                    _initialLat = lat;
+                    _initialLng = lng;
+                    _initialZoom = zoom;
+                  });
+                  _saveAndNotify();
+                },
               ),
-              if (!_rememberViewport) ...[
-                const SizedBox(height: 12),
-                InitialViewportSelector(
-                  lat: _initialLat,
-                  lng: _initialLng,
-                  zoom: _initialZoom,
-                  onChanged: (lat, lng, zoom) {
-                    setState(() {
-                      _initialLat = lat;
-                      _initialLng = lng;
-                      _initialZoom = zoom;
-                    });
-                    _saveAndNotify();
-                  },
-                ),
-              ],
               const Divider(height: 24),
 
-              // Map Source Selection
-              const Text(
-                'Map Source',
-                style: TextStyle(
-                  fontSize: 14,
-                  fontWeight: FontWeight.bold,
-                  color: Colors.grey,
-                ),
-              ),
-              const SizedBox(height: 12),
-              Container(
-                padding: const EdgeInsets.symmetric(
-                  horizontal: 12,
-                  vertical: 4,
-                ),
-                decoration: BoxDecoration(
-                  border: Border.all(color: Colors.grey.shade300),
-                  borderRadius: BorderRadius.circular(12),
-                ),
-                child: DropdownButton<MapSource>(
-                  value: _mapSource,
-                  isExpanded: true,
-                  underline: const SizedBox(),
-                  icon: const Icon(Icons.arrow_drop_down),
-                  items: MapSource.values.map((source) {
-                    return DropdownMenuItem<MapSource>(
-                      value: source,
-                      child: Row(
-                        children: [
-                          Icon(
-                            source.icon,
-                            size: 20,
-                            color: Colors.grey.shade700,
-                          ),
-                          const SizedBox(width: 12),
-                          Expanded(
-                            child: Column(
-                              crossAxisAlignment: CrossAxisAlignment.start,
-                              mainAxisSize: MainAxisSize.min,
-                              children: [
-                                Text(
-                                  source.displayName,
-                                  style: const TextStyle(
-                                    fontSize: 14,
-                                    fontWeight: FontWeight.w500,
-                                  ),
-                                ),
-                                Text(
-                                  source.description,
-                                  style: TextStyle(
-                                    fontSize: 11,
-                                    color: Colors.grey.shade600,
-                                  ),
-                                ),
-                              ],
-                            ),
-                          ),
-                        ],
-                      ),
-                    );
-                  }).toList(),
-                  onChanged: (MapSource? newValue) {
-                    if (newValue != null) {
-                      setState(() => _mapSource = newValue);
-                      _saveAndNotify();
-                    }
-                  },
-                ),
+              // Map source section
+              buildMapSourceSection(
+                mapSource: _mapSource,
+                onMapSourceChanged: (source) {
+                  setState(() => _mapSource = source);
+                  _saveAndNotify();
+                },
               ),
               const SizedBox(height: 12),
               const Divider(height: 24),
 
-              // Color customization.
-              const Text(
-                'Marker Colors',
-                style: TextStyle(
-                  fontSize: 14,
-                  fontWeight: FontWeight.bold,
-                  color: Colors.grey,
-                ),
-              ),
-              const SizedBox(height: 12),
-
-              // User places color.
-              ColorPickerTile(
-                label: 'My Places',
-                subtitle: 'Your saved locations',
-                color: _userPlacesColor,
-                onTap: () => _showColorPicker(
-                  title: 'My Places Color',
-                  currentColor: _userPlacesColor,
-                  onColorChanged: (color) {
-                    setState(() => _userPlacesColor = color);
-                    _saveAndNotify();
-                  },
-                ),
-              ),
-              const SizedBox(height: 12),
-
-              // Local places color.
-              ColorPickerTile(
-                label: 'Example Places',
-                subtitle: 'Canned example locations',
-                color: _localPlacesColor,
-                onTap: () => _showColorPicker(
-                  title: 'Example Places Color',
-                  currentColor: _localPlacesColor,
-                  onColorChanged: (color) {
-                    setState(() => _localPlacesColor = color);
-                    _saveAndNotify();
-                  },
-                ),
+              // Marker colors section
+              buildMarkerColorsSection(
+                context: context,
+                userPlacesColor: _userPlacesColor,
+                localPlacesColor: _localPlacesColor,
+                onUserColorChanged: (color) {
+                  setState(() => _userPlacesColor = color);
+                  _saveAndNotify();
+                },
+                onLocalColorChanged: (color) {
+                  setState(() => _localPlacesColor = color);
+                  _saveAndNotify();
+                },
               ),
               const SizedBox(height: 20),
 
-              // Reset button.
-              Center(
-                child: TextButton.icon(
-                  onPressed: _resetToDefaults,
-                  icon: const Icon(Icons.restore, size: 18),
-                  label: const Text('Reset to Defaults'),
-                  style: TextButton.styleFrom(foregroundColor: Colors.grey),
-                ),
-              ),
-
+              // Reset button
+              buildResetButton(onReset: _resetToDefaults),
               const SizedBox(height: 12),
 
-              // Logout button - only show if user is logged in
-              FutureBuilder<String?>(
-                future: getWebId(),
-                builder: (context, snapshot) {
-                  final isLoggedIn =
-                      snapshot.data != null && snapshot.data!.isNotEmpty;
-                  if (!isLoggedIn) return const SizedBox.shrink();
-
-                  return Column(
-                    children: [
-                      Center(
-                        child: TextButton.icon(
-                          onPressed: () async {
-                            // Close settings dialog first
-                            Navigator.pop(context);
-                            // Then handle logout
-                            await SolidAuthHandler.instance.handleLogout(
-                              context,
-                            );
-                          },
-                          icon: const Icon(Icons.logout, size: 18),
-                          label: const Text('Logout'),
-                          style: TextButton.styleFrom(
-                            foregroundColor: Colors.red.shade400,
-                          ),
-                        ),
-                      ),
-                      const SizedBox(height: 8),
-                      // DEBUG: Delete encryption keys from server
-                      Center(
-                        child: TextButton.icon(
-                          onPressed: () => _deleteEncryptionKeys(context),
-                          icon: const Icon(Icons.delete_forever, size: 18),
-                          label: const Text('Delete Encryption Keys (DEBUG)'),
-                          style: TextButton.styleFrom(
-                            foregroundColor: Colors.orange.shade700,
-                          ),
-                        ),
-                      ),
-                    ],
-                  );
-                },
-              ),
+              // User actions (logout, debug buttons)
+              buildUserActionsSection(context),
             ],
           ),
         ),
